fix set local race error & refactor
