+++ internal/runner/runner_test.go
@@ -18,10 +18,13 @@ package runner
 
 import (
 	"context"
+	stderrs "errors"
 	"os"
+	"syscall"
 	"testing"
 	"time"
 
+	"github.com/vdaas/vald/internal/config"
 	"github.com/vdaas/vald/internal/errors"
 	"github.com/vdaas/vald/internal/log"
 	"github.com/vdaas/vald/internal/log/logger"
@@ -303,47 +306,184 @@ func TestDo(t *testing.T) {
 		}
 		return nil
 	}
+	defaultAfterFunc := func(args) {
+		os.Args = []string{
+			"test",
+		}
+	}
 	tests := []test{
-		// TODO test cases
-		/*
-		   {
-		       name: "test_case_1",
-		       args: args {
-		           ctx: nil,
-		           opts: nil,
-		       },
-		       want: want{},
-		       checkFunc: defaultCheckFunc,
-		   },
-		*/
+		{
+			name: "returns nil when option is nil and version option is set",
+			args: args{
+				ctx: context.Background(),
+			},
+			beforeFunc: func(args) {
+				os.Args = []string{
+					"test", "-version",
+				}
+			},
+			want: want{
+				err: nil,
+			},
+		},
 
-		// TODO test cases
-		/*
-		   func() test {
-		       return test {
-		           name: "test_case_2",
-		           args: args {
-		           ctx: nil,
-		           opts: nil,
-		           },
-		           want: want{},
-		           checkFunc: defaultCheckFunc,
-		       }
-		   }(),
-		*/
+		{
+			name: "returns error when option is nil and params.Parse returns error",
+			args: args{
+				ctx: context.Background(),
+			},
+			beforeFunc: func(args) {
+				os.Args = []string{
+					"test", "-team=set",
+				}
+			},
+			want: want{
+				err: errors.ErrArgumentParseFailed(stderrs.New("flag provided but not defined: -team")),
+			},
+		},
+
+		{
+			name: "returns error when option is not nil and r.loadConfig returns error",
+			args: args{
+				ctx: context.Background(),
+				opts: []Option{
+					WithConfigLoader(func(string) (interface{}, *config.GlobalConfig, error) {
+						return nil, nil, errors.New("err")
+					}),
+				},
+			},
+			beforeFunc: func(args) {
+				os.Args = []string{
+					"test", "-c=./runner.go",
+				}
+			},
+			want: want{
+				err: errors.New("err"),
+			},
+		},
+
+		{
+			name: "returns error when option is not nil and ver.Check returns error",
+			args: args{
+				ctx: context.Background(),
+				opts: []Option{
+					WithVersion("v1.1.7", "v1.1.5", "v1.1.0"),
+					WithConfigLoader(func(string) (interface{}, *config.GlobalConfig, error) {
+						return nil, &config.GlobalConfig{
+							Logging: &config.Logging{
+								Logger: "glg",
+								Level:  "info",
+								Format: "json",
+							},
+							Version: "v1.1.7",
+						}, nil
+					}),
+				},
+			},
+			beforeFunc: func(args) {
+				os.Args = []string{
+					"test", "-c=./runner.go",
+				}
+			},
+			want: want{
+				err: errors.ErrInvalidConfigVersion("1.1.7", ">= v1.1.0, <= v1.1.5"),
+			},
+		},
+
+		{
+			name: "returns error when option is not nil and r.initializeDaemon returns error",
+			args: args{
+				ctx: context.Background(),
+				opts: []Option{
+					WithVersion("v1.1.2", "v1.1.5", "v1.1.0"),
+					WithConfigLoader(func(string) (interface{}, *config.GlobalConfig, error) {
+						return nil, &config.GlobalConfig{
+							Logging: &config.Logging{
+								Logger: "glg",
+								Level:  "info",
+								Format: "json",
+							},
+							Version: "v1.1.2",
+						}, nil
+					}),
+					WithDaemonInitializer(func(interface{}) (Runner, error) {
+						return nil, errors.New("err")
+					}),
+				},
+			},
+			beforeFunc: func(args) {
+				os.Args = []string{
+					"test", "-c=./runner.go",
+				}
+			},
+			want: want{
+				err: errors.New("err"),
+			},
+		},
+
+		{
+			name: "returns nil when option is not nil and Run returns nil",
+			args: args{
+				ctx: context.Background(),
+				opts: []Option{
+					WithVersion("v1.1.2", "v1.1.5", "v1.1.0"),
+					WithConfigLoader(func(string) (interface{}, *config.GlobalConfig, error) {
+						return nil, &config.GlobalConfig{
+							Logging: &config.Logging{
+								Logger: "glg",
+								Level:  "info",
+								Format: "json",
+							},
+							Version: "v1.1.2",
+						}, nil
+					}),
+					WithDaemonInitializer(func(interface{}) (Runner, error) {
+						return &runnerMock{
+							PreStartFunc: func(ctx context.Context) error {
+								return nil
+							},
+							StartFunc: func(ctx context.Context) (<-chan error, error) {
+								return make(chan error, 1), nil
+							},
+							PreStopFunc: func(ctx context.Context) error {
+								return nil
+							},
+							StopFunc: func(ctx context.Context) error {
+								return nil
+							},
+							PostStopFunc: func(ctx context.Context) error {
+								return nil
+							},
+						}, nil
+					}),
+				},
+			},
+			beforeFunc: func(args) {
+				os.Args = []string{
+					"test", "-c=./runner.go",
+				}
+				go func() {
+					time.Sleep(2 * time.Second)
+					syscall.Kill(syscall.Getpid(), syscall.SIGINT)
+				}()
+			},
+			want: want{
+				err: nil,
+			},
+		},
 	}
 
-	for _, tc := range tests {
-		test := tc
+	for _, test := range tests {
 		t.Run(test.name, func(tt *testing.T) {
-			tt.Parallel()
-			defer goleak.VerifyNone(tt, goleak.IgnoreCurrent())
+			defer goleak.VerifyNone(tt, goleakIgnoreOptions...)
 			if test.beforeFunc != nil {
 				test.beforeFunc(test.args)
 			}
-			if test.afterFunc != nil {
-				defer test.afterFunc(test.args)
+			if test.afterFunc == nil {
+				test.afterFunc = defaultAfterFunc
 			}
+			defer test.afterFunc(test.args)
+
 			checkFunc := test.checkFunc
 			if test.checkFunc == nil {
 				checkFunc = defaultCheckFunc
