+++ internal/timeutil/location/loc_test.go
@@ -23,6 +23,143 @@ import (
 	"github.com/vdaas/vald/internal/errors"
 )
 
+func TestSet(t *testing.T) {
+	type test struct {
+		name      string
+		loc       string
+		checkFunc func(got *time.Location) error
+	}
+
+	tests := []test{
+		{
+			name: "returns UTC location when loc is UTC",
+			loc:  locationUTC,
+			checkFunc: func(got *time.Location) error {
+				if !reflect.DeepEqual(got, time.UTC) {
+					return errors.Errorf("not equals. want: %v, but got: %v", time.UTC, got)
+				}
+				return nil
+			},
+		},
+
+		{
+			name: "returns UTC location when loc is Utc",
+			loc:  "UTc",
+			checkFunc: func(got *time.Location) error {
+				if !reflect.DeepEqual(got, time.UTC) {
+					return errors.Errorf("not equals. want: %v, but got: %v", time.UTC, got)
+				}
+				return nil
+			},
+		},
+
+		{
+			name: "returns GMT location when loc is GMT",
+			loc:  locationGMT,
+			checkFunc: func(got *time.Location) error {
+				if got == nil {
+					return errors.New("got is nil")
+				} else if got, want := got.String(), locationGMT; got != want {
+					return errors.Errorf("String() not equals. want: %v, but got: %v", want, got)
+				}
+				return nil
+			},
+		},
+
+		{
+			name: "returns GMT location when loc is Gmt",
+			loc:  "Gmt",
+			checkFunc: func(got *time.Location) error {
+				if got == nil {
+					return errors.New("got is nil")
+				} else if got, want := got.String(), locationGMT; got != want {
+					return errors.Errorf("String() not equals. want: %v, but got: %v", want, got)
+				}
+				return nil
+			},
+		},
+
+		{
+			name: "returns JST location when loc is JST",
+			loc:  locationJST,
+			checkFunc: func(got *time.Location) error {
+				if got == nil {
+					return errors.New("got is nil")
+				} else if got, want := got.String(), locationJST; got != want {
+					return errors.Errorf("String() not equals. want: %v, but got: %v", want, got)
+				}
+				return nil
+			},
+		},
+
+		{
+			name: "returns JST location when loc is Jst",
+			loc:  "Jst",
+			checkFunc: func(got *time.Location) error {
+				if got == nil {
+					return errors.New("got is nil")
+				} else if got, want := got.String(), locationJST; got != want {
+					return errors.Errorf("String() not equals. want: %v, but got: %v", want, got)
+				}
+				return nil
+			},
+		},
+
+		{
+			name: "returns JST location when loc is Asia/Tokyo",
+			loc:  locationTokyo,
+			checkFunc: func(got *time.Location) error {
+				if got == nil {
+					return errors.New("got is nil")
+				} else if got, want := got.String(), locationJST; got != want {
+					return errors.Errorf("String() not equals. want: %v, but got: %v", want, got)
+				}
+				return nil
+			},
+		},
+
+		{
+			name: "returns JST location when loc is ASIA/Tokyo",
+			loc:  "ASIA/Tokyo",
+			checkFunc: func(got *time.Location) error {
+				if got == nil {
+					return errors.New("got is nil")
+				} else if got, want := got.String(), locationJST; got != want {
+					return errors.Errorf("String() not equals. want: %v, but got: %v", want, got)
+				}
+				return nil
+			},
+		},
+
+		{
+			name: "returns invalid location when loc is invalid",
+			loc:  "invalid",
+			checkFunc: func(got *time.Location) error {
+				if got == nil {
+					return errors.New("got is nil")
+				} else if got, want := got.String(), "invalid"; got != want {
+					return errors.Errorf("String() not equals. want: %v, but got: %v", want, got)
+				}
+				return nil
+			},
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			time.Local = nil
+			defer func() {
+				time.Local = nil
+			}()
+			Set(tt.loc)
+			got := time.Local
+			if err := tt.checkFunc(got); err != nil {
+				t.Error(err)
+			}
+		})
+	}
+}
+
 func TestGMT(t *testing.T) {
 	tests := []struct {
 		name string
